Added Level 2 and sorting layers to scene 1
Most of the changes in this branch are the addition of the second scene
which also uses the Global Behavior script. I also added sorting layers
to ShaneleeTran_mp3. I also added a background sprite to scene one and a
changed a few textures, but I expect shane will change those again.